requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-login==0.6.2
Werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1 